Move LPGitFetcher upstream and add missing rename (fixes #14)
