+++ charmtools/fetchers.py
@@ -117,6 +117,21 @@ class BzrMergeProposalFetcher(BzrFetcher):
         return rename(dir_)
 
 
+class LaunchpadGitFetcher(Fetcher):
+    MATCH = re.compile(r"""
+    ^(git:|https)?://git.launchpad.net/
+    (?P<repo>[^@]*)(@(?P<revision>.*))?$
+    """, re.VERBOSE)
+
+    def fetch(self, dir_):
+        dir_ = tempfile.mkdtemp(dir=dir_)
+        url = 'https://git.launchpad.net/' + self.repo
+        git('clone {} {}'.format(url, dir_))
+        if self.revision:
+            git('checkout {}'.format(self.revision), cwd=dir_)
+        return rename(dir_)
+
+
 class GithubFetcher(Fetcher):
     MATCH = re.compile(r"""
     ^(gh:|github:|https?://(www\.)?github.com/)
@@ -314,6 +329,7 @@ FETCHERS = [
     LocalFetcher,
     CharmstoreDownloader,
     BundleDownloader,
+    LaunchpadGitFetcher,
 ]
